tpm: Get rid of AbstractTPM class
AbstractTPM may have served as an interface-defining class while TPM 1.2
support was still there but now it's obsolete, so remove it.
